requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
